Refactor: Moved cache functions of setup.py to cache.py: renamed setup.py to config.py
